+++ crates/pixi-build-backend/src/traits/mod.rs
@@ -1,0 +1,7 @@
+pub mod package_spec;
+pub mod project;
+pub mod targets;
+
+pub use package_spec::{AnyVersion, BinarySpecExt, PackageSpec};
+pub use project::ProjectModel;
+pub use targets::{Dependencies, TargetSelector, Targets};
